--- tests/baseline_result.json
@@ -1,1 +1,0 @@
-{"categorical": ["Code_region", "Apprentissage", "Segpa", "Section_cinema", "Section_theatre", "Section_internationale", "Lycee_Agricole", "Lycee_militaire", "Lycee_des_metiers", "Libelle_region", "epsg", "precision_localisation", "date_maj_ligne", "etat", "ministere_tutelle", "etablissement_multi_lignes", "rpi_concentre"], "columns": {"Identifiant_de_l_etablissement": ["uai"], "Adresse_1": ["adresse"], "Code postal": ["code_postal"], "Code_commune": ["code_commune_insee"], "Nom_commune": ["commune"], "Code_departement": ["code_departement"], "Code_academie": ["code_departement"], "Code_region": ["code_departement", "code_region"], "Ecole_maternelle": ["booleen"], "Ecole_elementaire": ["booleen"], "Voie_generale": ["booleen"], "Voie_technologique": ["booleen"], "Voie_professionnelle": ["booleen"], "Fax": ["tel_fr"], "Web": ["url"], "Mail": ["email"], "Restauration": ["booleen"], "Hebergement": ["booleen"], "ULIS": ["booleen"], "Apprentissage": ["booleen"], "Segpa": ["booleen"], "Section_arts": ["booleen"], "Section_cinema": ["booleen"], "Section_theatre": ["booleen"], "Section_sport": ["booleen"], "Section_internationale": ["booleen"], "Section_europeenne": ["booleen"], "Lycee_Agricole": ["booleen"], "Lycee_militaire": ["booleen"], "Lycee_des_metiers": ["booleen"], "Post_BAC": ["booleen"], "GRETA": ["booleen"], "SIREN_SIRET": ["siret"], "Fiche_onisep": ["url"], "position": ["latlon_wgs"], "Libelle_departement": ["departement"], "Libelle_academie": ["commune"], "Libelle_region": ["region"], "latitude": ["latitude_wgs", "longitude_wgs", "longitude_wgs_fr_metropole", "longitude_l93"], "longitude": ["latitude_wgs", "longitude_wgs", "latitude_wgs_fr_metropole", "longitude_l93"], "precision_localisation": ["adresse"], "date_ouverture": ["date"], "date_maj_ligne": ["date"], "etablissement_multi_lignes": ["booleen"], "rpi_concentre": ["booleen"], "PIAL": ["uai"]}, "continous": ["coordonnee_X", "coordonnee_Y", "latitude", "longitude"], "encoding": "UTF-8", "header": ["Identifiant_de_l_etablissement", "Nom_etablissement", "Type_etablissement", "Statut_public_prive", "Adresse_1", "Adresse_2", "Adresse_3", "Code postal", "Code_commune", "Nom_commune", "Code_departement", "Code_academie", "Code_region", "Ecole_maternelle", "Ecole_elementaire", "Voie_generale", "Voie_technologique", "Voie_professionnelle", "Telephone", "Fax", "Web", "Mail", "Restauration", "Hebergement", "ULIS", "Apprentissage", "Segpa", "Section_arts", "Section_cinema", "Section_theatre", "Section_sport", "Section_internationale", "Section_europeenne", "Lycee_Agricole", "Lycee_militaire", "Lycee_des_metiers", "Post_BAC", "Appartenance_Education_Prioritaire", "GRETA", "SIREN_SIRET", "Nombre_d_eleves", "Fiche_onisep", "position", "Type_contrat_prive", "Libelle_departement", "Libelle_academie", "Libelle_region", "coordonnee_X", "coordonnee_Y", "epsg", "nom_circonscription", "latitude", "longitude", "precision_localisation", "date_ouverture", "date_maj_ligne", "etat", "ministere_tutelle", "etablissement_multi_lignes", "rpi_concentre", "rpi_disperse", "code_nature", "libelle_nature", "Code_type_contrat_prive", "PIAL"], "header_row_idx": 0, "heading_columns": 0, "ints_as_floats": [], "separator": ";", "total_lines": 9399, "trailing_columns": 0}
